+++ src/main/java/org/openrewrite/java/spring/security5/UseNewRequestMatchers.java
@@ -27,6 +27,7 @@ import org.openrewrite.java.JavaTemplate;
 import org.openrewrite.java.MethodMatcher;
 import org.openrewrite.java.search.UsesMethod;
 import org.openrewrite.java.tree.J;
+import org.openrewrite.java.tree.JavaType;
 
 import static java.util.stream.Collectors.joining;
 
@@ -39,7 +40,6 @@ public class UseNewRequestMatchers extends Recipe {
     private static final MethodMatcher MVC_MATCHERS = new MethodMatcher(CLAZZ + " mvcMatchers(..)", true);
     private static final MethodMatcher REGEX_MATCHERS = new MethodMatcher(CLAZZ + " regexMatchers(..)");
     private static final MethodMatcher CSRF_MATCHERS = new MethodMatcher("org.springframework.security.config.annotation.web.configurers.CsrfConfigurer ignoringAntMatchers(..)");
-
 
     @Override
     public String getDisplayName() {
@@ -73,7 +73,7 @@ public class UseNewRequestMatchers extends Recipe {
                                     .build();
                             J.MethodInvocation apply = template.apply(getCursor(), mi.getCoordinates().replaceMethod(), mi.getArguments().toArray());
                             return apply.withSelect(mi.getSelect())
-                                    .withName(mi.getName().withSimpleName(replacementMethodName));
+                                    .withName(mi.getName().withSimpleName(replacementMethodName).withType(JavaType.buildType(CLAZZ)));
                         }
                         return mi;
                     }
